requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
databricks-sdk>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
